adapters/dominion-observatory-autogen/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "dominion-observatory-autogen"
version = "0.1.0"
description = "Microsoft AutoGen (v0.4+) integration for the Dominion Observatory — instrument autogen_core.tools.BaseTool calls with anonymised runtime telemetry to the cross-ecosystem MCP trust layer."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.14"
authors = [
    { name = "Dominion Observatory", email = "hello@levylens.co" },
]
keywords = ["autogen", "autogen-core", "mcp", "observability", "trust", "telemetry", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "dominion-observatory-sdk>=0.2.0",
]

[project.optional-dependencies]
autogen = ["autogen-core>=0.4.0"]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://dominion-observatory.sgdata.workers.dev"
Repository = "https://github.com/vdineshk/daee-hitman"
Issues = "https://github.com/vdineshk/daee-hitman/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/dominion_observatory_autogen"]

[tool.hatch.build.targets.sdist]
include = [
    "src/dominion_observatory_autogen",
    "README.md",
    "LICENSE",
]
